Fix what the three-OS matrix caught on its first run
The matrix went red on both new legs the first time it ran, which is the
whole reason it exists. One of the failures is a defect in shipped code.

internal/strategy: the credit gate answers wrongly on arm64

    `spendArmFraction*capacity - *s.Used` is a shape Go permits a compiler to
    fuse into one FMA -- explicitly "possibly across statements", so an
    intermediate variable does not stop it -- and arm64 does. Fused, the
    exact product 0.90*100 (90.000000000000002220446…) is never rounded to 90
    before the subtraction, so $90 spent against a $100 cap answers 2.22e-15
    of room and `room > 0` is TRUE. §7.3 says every branch of this function
    fails closed because it is the only place in ccdad where a bug spends the
    user's money; on darwin/arm64 and linux/arm64, two of the six shipped
    targets, one branch did not.

    An explicit float64 conversion is what the spec says forbids the fusion.
    Verified in the generated code rather than by reasoning: credit.go
    compiled FNMSUBD for GOARCH=arm64 before and compiles FMULD + FSUBD
    after. A sweep of `go build -gcflags=all=-S` for arm64 over every package
    finds exactly one other contraction, in pollpolicy's jitter(), where a
    one-ulp difference in a deliberately randomised interval decides nothing.

Windows has no mode bits

    os.Chmod there toggles the read-only attribute and nothing else, and Stat
    reports 0666 whatever the file was created with. §10.3 already accepts
    that for v1 and relies on the inherited %USERPROFILE% ACL, which is a
    property of the directory rather than of the file, so the four `perm !=
    0600` assertions now say so instead of failing.

    The same no-op is why two config tests and one cclock test failed: they
    make a file unreadable, or a directory unremovable, with chmod and then
    assert the failure that causes. Chmod REPORTS success on Windows, so the
    premise quietly evaporated. They now probe -- read the file, create the
    directory -- and skip when the platform ignored the mode. The probe also
    subsumes the os.Geteuid() == 0 guards, which could never have covered
    this: Geteuid answers -1 on Windows.

ENOLCK is a unix errno

    locksUnsupported has no errno cases off unix, by design and with a
    comment saying so, so injecting syscall.ENOLCK on Windows asserted that
    classifyLockError does something it was never asked to. The daemon and
    store tests now inject the error the platform actually reports -- ENOLCK
    on unix, errors.ErrUnsupported elsewhere -- and keep their meaning on
    every leg. ENOSYS needed nothing: syscall.Errno.Is maps it to
    errors.ErrUnsupported everywhere, which is why that row passed on Windows
    while its neighbour failed.

A reader with no pause starves the replace on Windows

    FILE_SHARE_DELETE stops the torn-file detector BLOCKING the rename, but
    the replaced file then sits delete-pending until the handle closes, and
    MoveFileEx onto a name in that state answers ERROR_ACCESS_DENIED. Reopened
    with no pause at all, the name never leaves that state and the writer
    spends all ten of the bounded retries §10.3 sizes for a transient
    antivirus handle. One millisecond between reads on Windows only; off it a
    rename cannot be obstructed by an open handle and the tightest possible
    loop is the best detector.

+++ internal/strategy/credit.go
@@ -52,7 +52,21 @@ func CreditRoom(s SpendInfo, ceiling float64) (float64, bool) {
 	if s.Used == nil {
 		return 0, false // wire drift must not hand back the full cap
 	}
-	room := spendArmFraction*capacity - *s.Used
+	// The float64 conversion is load-bearing, not decoration. Go permits a
+	// compiler to fuse `a*b - c` into a single FMA -- possibly ACROSS
+	// statements, so an intermediate variable does not stop it -- and arm64
+	// does. Fused, the exact product 0.90*100 (90.000000000000002220446…) is
+	// never rounded to 90 before the subtraction, so $90 spent against a $100
+	// cap answers 2.22e-15 of room, and `room > 0` is TRUE. That is not a
+	// display artefact: it is the credit gate reporting room on the one axis
+	// in ccdad where a wrong answer spends the user's money, and darwin/arm64
+	// and linux/arm64 are two of the six shipped targets.
+	//
+	// An explicit conversion is the one thing the spec says forbids the
+	// fusion, which makes all six targets agree. Measured: green on amd64,
+	// red on macos-latest (arm64), where CreditRoom(limit 100, used 80) came
+	// back as 10.000000000000002.
+	room := float64(spendArmFraction*capacity) - *s.Used
 	// A NaN anywhere in the account's own figures reaches here as a NaN room,
 	// and `room > 0` is false for it — which is the fail-closed answer.
 	return room, room > 0
